Delegate to schema to check compatibility of insertable content
