+++ packages/dds/tree/src/treeFactory.ts
@@ -100,7 +100,6 @@ export const SharedTree = configuredSharedTree({});
  * 	TreeCompressionStrategy,
  * 	configuredSharedTree,
  * 	typeboxValidator,
- * 	// eslint-disable-next-line import/no-internal-modules
  * } from "@fluidframework/tree/internal";
  * const SharedTree = configuredSharedTree({
  * 	forest: ForestTypeReference,
